FTPRetriever
Recreating program from previous job. This script will check a preset folder, on a pre-determined timer (every 5 seconds, as of right now) and, if a .go file is present, will transfer the .zip equivalent to an FTP server.

The .go file is the trigger file, allowing the .zip file to remain in the folder until the end-user is ready for the file to send. 

Script to be updated with logging functionality.